backend/requirements.txt
uvicorn
sqlalchemy
fastapi
python-jose[cryptography]
python-multipart
requests
passlib
bcrypt
stripe